Renamed Project Back to Duke bacause I cant figure out how the runtest scripts work.
