+++ src/test/java/seedu/duke/DukeTest.java
@@ -1,4 +1,4 @@
-package seedu.situs;
+package seedu.duke;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
